Separate selectors in state/ui.
Moves each ui selector to its own module, to improve code-splitting in
future changes. To ease migration, there's still an index file which
imports and reexports them all.
