+++ client/state/ui/selectors/get-selected-site-id.js
@@ -1,0 +1,9 @@
+/**
+ * Returns the currently selected site ID.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {?Number}       Selected site ID
+ */
+export default function getSelectedSiteId( state ) {
+	return state.ui.selectedSiteId;
+}
