+++ client/state/ui/selectors/get-selected-site-slug.js
@@ -1,0 +1,21 @@
+/**
+ * Internal dependencies
+ */
+import { getSiteSlug } from 'state/sites/selectors';
+import getSelectedSiteId from './get-selected-site-id';
+
+/**
+ * Returns the slug of the currently selected site,
+ * or null if no site is selected.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {?String}       Selected site slug
+ */
+export default function getSelectedSiteSlug( state ) {
+	const siteId = getSelectedSiteId( state );
+	if ( ! siteId ) {
+		return null;
+	}
+
+	return getSiteSlug( state, siteId );
+}
