+++ client/state/ui/selectors/get-selected-site.js
@@ -1,0 +1,20 @@
+/**
+ * Internal dependencies
+ */
+import { getSite } from 'state/sites/selectors';
+import getSelectedSiteId from './get-selected-site-id';
+
+/**
+ * Returns the site object for the currently selected site.
+ *
+ * @param  {Object}  state  Global state tree
+ * @return {?Object}        Selected site
+ */
+export default function getSelectedSite( state ) {
+	const siteId = getSelectedSiteId( state );
+	if ( ! siteId ) {
+		return null;
+	}
+
+	return getSite( state, siteId );
+}
