+++ client/state/ui/selectors/is-preview-showing.js
@@ -1,0 +1,16 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Returns true if WebPreview is currently showing.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {bool}    True if currently showing WebPreview
+ *
+ * @see client/components/web-preview
+ */
+export default function isPreviewShowing( state ) {
+	return get( state.ui, 'isPreviewShowing', false );
+}
