+++ client/state/ui/selectors/is-section-isomorphic.js
@@ -1,0 +1,21 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import getSection from './get-section';
+
+/**
+ * Returns true if the current section is isomorphic.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {bool}    True if current section is isomorphic
+ *
+ * @see client/sections
+ */
+export default function isSectionIsomorphic( state ) {
+	return get( getSection( state ), 'isomorphic', false );
+}
