+++ client/state/ui/selectors/is-section-loading.js
@@ -1,0 +1,9 @@
+/**
+ * Returns whether a section is loading.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {Boolean}       Whether the section is loading
+ */
+export default function isSectionLoading( state ) {
+	return state.ui.isLoading;
+}
