+++ client/state/ui/selectors/masterbar-is-visible.js
@@ -1,0 +1,3 @@
+export default function masterbarIsVisible( state ) {
+	return state.ui.masterbarVisibility;
+}
